+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import json
+import logging
 from base64 import b64encode
 from collections import deque
 from functools import partial
@@ -9,17 +10,17 @@ from unittest.mock import MagicMock, Mock
 import pytest
 from fastapi.testclient import TestClient
 from neo4j import Driver, Session, SummaryCounters, Transaction
-from pytest import FixtureRequest, MonkeyPatch
+from pytest import FixtureRequest, LogCaptureFixture, MonkeyPatch
 from valkey import Valkey
 
 from mex.artificial.helpers import create_artificial_items_and_rule_sets
 from mex.backend.cache.connector import CacheConnector, LocalCache, ValkeyCache
 from mex.backend.graph.connector import GraphConnector
 from mex.backend.main import app
-from mex.backend.models import APIKeyDatabase, APIUserDatabase
 from mex.backend.settings import BackendSettings
 from mex.backend.testing.main import app as testing_app
 from mex.common.connector import CONNECTOR_STORE
+from mex.common.logging import logger
 from mex.common.models import (
     MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
     AdditiveOrganizationalUnit,
@@ -48,16 +49,64 @@ pytest_plugins = ("mex.common.testing.plugin",)
 
 
 @pytest.fixture(autouse=True)
-def settings() -> BackendSettings:
+def settings(
+    caplog: LogCaptureFixture,
+    monkeypatch: MonkeyPatch,
+    is_integration_test: bool,  # noqa: FBT001
+    log_level: int,
+    isolate_settings: None,  # noqa: ARG001
+) -> BackendSettings:
     """Load the settings for this pytest session."""
-    settings = BackendSettings.get()
-    settings.backend_api_key_database = APIKeyDatabase(
-        read=["read_key"], write=["write_key"]
+    monkeypatch.setenv(
+        "MEX_BACKEND_API_KEY_DATABASE",
+        json.dumps(
+            {
+                "read": ["read_key"],
+                "write": ["write_key"],
+            }
+        ),
     )
-    settings.backend_user_database = APIUserDatabase(
-        read={"Reader": "read_password"}, write={"Writer": "write_password"}
+    monkeypatch.setenv(
+        "MEX_BACKEND_API_USER_DATABASE",
+        json.dumps(
+            {
+                "read": {"Reader": "read_password"},
+                "write": {"Writer": "write_password"},
+            }
+        ),
     )
-    return settings
+    if is_integration_test:
+        monkeypatch.setenv(
+            "MEX_IDENTITY_PROVIDER",
+            IdentityProvider.GRAPH.value,
+        )
+    else:
+        monkeypatch.setenv(
+            "MEX_IDENTITY_PROVIDER",
+            IdentityProvider.MEMORY.value,
+        )
+    # temporarily reduce log-level because the settings emit their configuration
+    # on every instantiation or value-change. this would flood the test logs with noise,
+    # especially because this fixture is used by *every* test.
+    with caplog.at_level(log_level, logger=logger.name):
+        return BackendSettings.get()
+
+
+@pytest.fixture
+def log_level(request: FixtureRequest) -> int:
+    """Returns a sensible log-level for the current pytest verbosity.
+
+    This can be controlled by adding more "v"s to `pytest -v`.
+    """
+    levels_by_verbosity = {
+        0: logging.ERROR,  # always shown
+        1: logging.WARNING,  # -v
+        2: logging.INFO,  # -vv
+    }
+    return levels_by_verbosity.get(
+        request.config.option.verbose,
+        logging.DEBUG,  # `-vvv` and above
+    )
 
 
 @pytest.fixture
@@ -201,21 +250,6 @@ def isolate_identifier_seeds(monkeypatch: MonkeyPatch) -> None:
 
 
 @pytest.fixture(autouse=True)
-def set_identity_provider(
-    is_integration_test: bool,  # noqa: FBT001
-    monkeypatch: MonkeyPatch,
-) -> None:
-    """Ensure the identifier provider is set correctly for unit and int tests."""
-    # TODO(ND): yuck, all this needs cleaning up after MX-1596
-    settings = BackendSettings.get()
-    if is_integration_test:
-        monkeypatch.setitem(settings.model_config, "validate_assignment", False)  # noqa: FBT003
-        monkeypatch.setattr(settings, "identity_provider", IdentityProvider.GRAPH)
-    else:
-        monkeypatch.setattr(settings, "identity_provider", IdentityProvider.MEMORY)
-
-
-@pytest.fixture(autouse=True)
 def isolate_graph_database(
     is_integration_test: bool,  # noqa: FBT001
     settings: BackendSettings,
@@ -303,7 +337,7 @@ DummyDataName = Literal[
 
 @pytest.fixture
 def dummy_data(
-    set_identity_provider: None,  # noqa: ARG001
+    settings: BackendSettings,  # noqa: ARG001
 ) -> DummyData:
     """Create interlinked dummy data and return a lookup by memorable names."""
     primary_source_1 = ExtractedPrimarySource(